+++ application/controllers/Welcome.php
@@ -9,6 +9,24 @@ class Welcome extends Application
 	{
 		parent::__construct();
 	}
+        
+        /**
+         * renders the page, selecting a random avatar and quote
+         */
+        public function random() 
+        {
+                // this is the view we want shown
+		$this->data['pagebody'] = 'homepage';
+
+		// build the list of authors, to pass on to our view
+		$source = $this->quotes->all();
+		$authors = array ();
+                $authors[] = $source[rand(0, count($source) - 1)];
+
+		$this->data['authors'] = $authors;
+
+		$this->render();
+        }
 
 	/**
 	 * Homepage for our app